# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.DefaultResponse:
    properties:
      message:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Company:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.SwagCompany:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8181
info:
  contact: {}
  description: This is a Tajik Career Hub API documentation.
  title: Tajik Career Hub API ✨
  version: "1.0"
paths:
...
  /companies:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all companies. No authentication required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Company'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all companies
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Add a new company to the database. Requires authentication.
      parameters:
      - description: Company data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/models.SwagCompany'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a new company
      tags:
      - Companies
  /companies/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a company by its ID. Requires authentication.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a company
      tags:
      - Companies
    get:
      consumes:
      - application/json
      description: Retrieve a single company by its ID. No authentication required.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get company by ID
      tags:
      - Companies
    put:
      consumes:
      - application/json
      description: Update a company by its ID. Requires authentication.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated company data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/models.SwagCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing company
      tags:
      - Companies
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"